move back to upstream esphome::RingBuffer which now supports avoiding partial writing
