+++ config/common/components.external.yaml
@@ -11,7 +11,6 @@ external_components:
         - fusb302b
         - i2s_audio
         - memory_flasher
-        - microphone
         - micro_wake_word
         - mixer
         - pcm5122
@@ -21,5 +20,4 @@ external_components:
         - snapcast
         - speaker
         - tas2780
-        - voice_assistant
