+++ config/satellite1.yaml
@@ -111,7 +111,6 @@ external_components:
         - fusb302b
         - i2s_audio
         - memory_flasher
-        - microphone
         - micro_wake_word
         - mixer
         - pcm5122
@@ -121,7 +120,6 @@ external_components:
         - snapcast
         - speaker
         - tas2780
-        - voice_assistant
