+++ esphome/components/audio/audio_transfer_buffer.h
@@ -4,7 +4,7 @@
 #include "chunked_ring_buffer.h"
 
 #include "esphome/core/defines.h"
-#include "ring_buffer.h"
+#include "esphome/core/ring_buffer.h"
 
 #ifdef USE_SPEAKER
 #include "esphome/components/speaker/speaker.h"
