+++ esphome/components/i2s_audio/speaker/i2s_audio_speaker.cpp
@@ -567,7 +567,7 @@ esp_err_t I2SAudioSpeaker::allocate_buffers_(size_t data_buffer_size, size_t rin
 
   if (this->audio_ring_buffer_.use_count() == 0) {
     // Allocate ring buffer. Uses a shared_ptr to ensure it isn't improperly deallocated.
-    this->audio_ring_buffer_ = audio::RingBuffer::create(ring_buffer_size);
+    this->audio_ring_buffer_ = RingBuffer::create(ring_buffer_size);
   }
 
   if (this->audio_ring_buffer_ == nullptr) {
